continued unit logic

+++ UnitConverter/UnitConverter_test.go
@@ -4,7 +4,7 @@ import "testing"
 
 func TestFahrenheitToCelsius(t *testing.T) {
 	result := ConvertUnit(32.0, "F", "C")
-	var wants float32 = 0.0
+	var wants float64 = 0.0
 	if result != wants {
 		t.Errorf("ConvertUnit(32.0, %q, %q) = %f, should be equal to %f", "F", "C", result, wants)
 	}
@@ -12,7 +12,7 @@ func TestFahrenheitToCelsius(t *testing.T) {
 
 func TestFahrenheitToKelvin(t *testing.T) {
 	result := ConvertUnit(44.0, "F", "K")
-	var wants float32 = 279.817
+	var wants float64 = 279.817
 	if result != wants {
 		t.Errorf("ConvertUnit(32.0, %q, %q) = %f, should be equal to %f", "F", "C", result, wants)
 	}
